MIFOS-4894: ensure validation for amount and interest is correct for too many digits before and after deciaml.

+++ userInterface/src/main/java/org/mifos/clientportfolio/loan/ui/LoanAccountController.java
@@ -286,7 +286,7 @@ public class LoanAccountController {
         loanScheduleFormBean.setActualPaymentDates(actualPaymentDates);
         loanScheduleFormBean.setActualPaymentAmounts(actualPaymentAmounts);
         
-        loanScheduleFormBean.setLoanPrincipal(BigDecimal.valueOf(formBean.getAmount()));
+        loanScheduleFormBean.setLoanPrincipal(BigDecimal.valueOf(formBean.getAmount().doubleValue()));
         loanScheduleFormBean.setTotalLoanInterest(totalLoanInterest);
         loanScheduleFormBean.setTotalLoanFees(totalLoanFees);
         loanScheduleFormBean.setRepaymentInstallments(loanSchedule.getInstallments());
